fix several CC warnings (taken from Debian)
git-svn-id: file:///tmp/cracklib-svn/trunk@133 4175fe1e-86d5-4fdc-8e6a-506fab9d8533

+++ cracklib/util/packer.c
@@ -17,8 +17,8 @@ main(argc, argv)
     int argc;
     char *argv[];
 {
-    uint32_t readed;
-    uint32_t wrote;
+    unsigned long readed;
+    unsigned long wrote;
     PWDICT *pwp;
     char buffer[STRINGSIZE];
     char *file;
